[Coronahack] Split adding labels to additional steps - Split adding labels and ids to separate steps - Moved dataset preparation to abstract function that is called only when source path exists - Enabled not-creating Masks folder if it is set to None (for classification-only datasets)

+++ src/steps/delete_imgs_with_no_annotations.py
@@ -16,7 +16,7 @@ class DeleteImgsWithNoAnnotations(TransformerMixin):
         mask_folder_name: str = "Masks",
         **kwargs: dict,
     ):
-        """Delete images without masks.
+        """Delete images without annotations.
 
         Args:
             mask_folder_name (str, optional): Name of the folder with masks. Defaults to "Masks".
@@ -24,7 +24,7 @@ class DeleteImgsWithNoAnnotations(TransformerMixin):
         self.mask_folder_name = mask_folder_name
 
     def transform(self, X: str) -> list:
-        """Delete images without masks.
+        """Delete images without annotations.
 
         Args:
             X (list): List of paths to the images.
@@ -32,9 +32,10 @@ class DeleteImgsWithNoAnnotations(TransformerMixin):
             X (list): List of paths to the images.
         """
         root_path = os.path.dirname(os.path.dirname(X[0]))
-        mask_paths = glob.glob(f"{os.path.join(root_path, self.mask_folder_name)}/**/*.png", recursive=True)
-        self.mask_names = [os.path.basename(mask) for mask in mask_paths]
-        print("Deleting images without masks...")
+        if self.mask_folder_name:
+            mask_paths = glob.glob(f"{os.path.join(root_path, self.mask_folder_name)}/**/*.png", recursive=True)
+            self.mask_names = [os.path.basename(mask) for mask in mask_paths]
+        print("Deleting images without annotations...")
 
         for img_path in tqdm(X):
             self.delete_imgs_with_no_annotations(img_path, root_path)
@@ -52,17 +53,19 @@ class DeleteImgsWithNoAnnotations(TransformerMixin):
             root_path (str): Path to the root folder.
         """
         img_name = os.path.basename(img_path)
-        mask_path = os.path.join(root_path, self.mask_folder_name, img_name)
+        if self.mask_folder_name:
+            mask_path = os.path.join(root_path, self.mask_folder_name, img_name)
         no_label = False
         if "-" not in img_name:
             no_label = True
-        if img_name not in self.mask_names:
-            if no_label:
-                # If there is no mask and no label
-                os.remove(img_path)
-        # If there is a mask but it is empty
-        elif np.unique(cv2.imread(mask_path)).shape[0] == 1:
-            if no_label:
-                # If there is a blank mask and no label
-                os.remove(img_path)
-                os.remove(mask_path)
+        if self.mask_folder_name:
+            if img_name not in self.mask_names:
+                if no_label:
+                    # If there is no mask and no label
+                    os.remove(img_path)
+            # If there is a mask but it is empty
+            elif np.unique(cv2.imread(mask_path)).shape[0] == 1:
+                if no_label:
+                    # If there is a blank mask and no label
+                    os.remove(img_path)
+                    os.remove(mask_path)
